Rename container with better names
