+++ src/containers/query-browser.jsx
@@ -5,7 +5,7 @@ import { Link } from 'react-router';
 import * as DatabaseActions from '../actions/databases.js';
 import * as QueryActions from '../actions/query.js';
 import DatabaseList from '../components/database-list.jsx';
-import Database from '../components/database.jsx';
+import Query from '../components/query.jsx';
 
 
 const STYLES = {
@@ -78,7 +78,7 @@ export default class DatabaseListContainer extends Component {
             </div>
           </div>
           <div style={STYLES.content}>
-            <Database query={query} actions={queryActions} />
+            <Query query={query} actions={queryActions} />
           </div>
         </div>
       </div>
